Add mover manager

+++ 3Dcity.XNA/3Dcity.XNA.SystemTests/Master/Managers/StopwatchManagerSystemTests.cs
@@ -22,7 +22,7 @@ namespace WindowsGame.SystemTests.Master.Managers
 			System.Threading.Thread.Sleep(2000);
 			StopwatchManager.Stop();
 			Int64 time = StopwatchManager.ElapsedMilliseconds;
-			Assert.That(2000, Is.EqualTo(time));
+			Assert.That((Int64)2000, Is.EqualTo(time));
 		}
 
 		[TearDown]
